feat(ptst): adiciona o operador  para verificar se uma expressão está em outro tipo de dado
